refactor: buat folder 'algoritma' dan pindahkan 'searching' ke dalamnya
