+++ algoritma/searching/interpolation.go
