+++ algoritma/searching/linear_searching.go
