+++ algoritma/searching/ternary_searching.go
