+++ algoritma/searching/utilities.go
