+++ tests/algoritma/searching/test_data_search.go
